Made Pacman move with a different method, but crashes.

+++ src/GhostObject.java
@@ -1,15 +1,15 @@
 import java.awt.Graphics;
 
 public class GhostObject {
-	int x;
-	int y;
+	double x;
+	double y;
 
-	public GhostObject(int x, int y) {
+	public GhostObject(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
 	public void draw(Graphics g) {
-		g.drawImage(GamePanel.ghostImg, x, y, MazeObject.blockWidth, MazeObject.blockHeight, null);
+		g.drawImage(GamePanel.ghostImg, (int)x, (int)y, MazeObject.blockWidth, MazeObject.blockHeight, null);
 	}
 }
